+++ adapter-out-mongodb/src/main/kotlin/de/chrgroth/spotify/control/adapter/out/mongodb/AppAlbumDocumentRepository.kt
@@ -1,0 +1,7 @@
+package de.chrgroth.spotify.control.adapter.out.mongodb
+
+import io.quarkus.mongodb.panache.kotlin.PanacheMongoRepositoryBase
+import jakarta.enterprise.context.ApplicationScoped
+
+@ApplicationScoped
+class AppAlbumDocumentRepository : PanacheMongoRepositoryBase<AppAlbumDocument, String>